+++ R/dataviz.R
@@ -88,3 +88,89 @@ theme_facet <- function(base_size = 12, base_family = "", border = TRUE, clean_x
   }
   return(theme)
 }
+
+#' @title Flat violin plot
+#' @description Violin plots are a compact display of continuous distributions
+#'   but are usually mirrored to mimick boxplots. The "flat" version removes
+#'   that mirrorness and makes the violin plots less...suggestive.
+#' @inheritParams ggplot2::geom_point
+#' @param trim If `TRUE` (default), trim the tails of the violins
+#'   to the range of the data. If `FALSE`, don't trim the tails.
+#' @param geom,stat Use to override the default connection between
+#'   `geom_violin` and `stat_ydensity`.
+#' @examples \dontrun{
+#' ggplot(diamonds, aes(cut, carat)) +
+#'   geom_flat_violin() +
+#'   coord_flip()
+#' }
+#' @author [David Robinson](https://github.com/dgrtwo)
+#' @source Gist: [dgrtwo/geom_flat_violin.R](https://gist.github.com/dgrtwo/eb7750e74997891d7c20)
+#' @rdname ggplot2-flatviolin
+#' @export
+geom_flat_violin <- function(mapping = NULL, data = NULL, stat = "ydensity",
+                             position = "dodge", trim = TRUE, scale = "area",
+                             show.legend = NA, inherit.aes = TRUE, ...) {
+  return(ggplot2::layer(
+    data = data,
+    mapping = mapping,
+    stat = stat,
+    geom = GeomFlatViolin,
+    position = position,
+    show.legend = show.legend,
+    inherit.aes = inherit.aes,
+    params = list(
+      trim = trim,
+      scale = scale,
+      ...
+    )
+  ))
+}
+
+"%||%" <- function(a, b) {
+  if (!is.null(a)) {
+    return(a)
+  } else {
+    return(b)
+  }
+}
+
+#' @rdname ggplot2-flatviolin
+#' @format NULL
+#' @usage NULL
+GeomFlatViolin <- ggplot2::ggproto(
+  "GeomFlatViolin",
+  ggplot2::Geom,
+  setup_data = function(data, params) {
+    data$width <- data$width %||%
+      params$width %||% (resolution(data$x, FALSE) * 0.9)
+
+    # ymin, ymax, xmin, and xmax define the bounding rectangle for each group
+    return(dplyr::mutate(
+      dplyr::group_by(data, group), ymin = min(y),
+      ymax = max(y),
+      xmin = x,
+      xmax = x + width / 2
+    ))
+  },
+
+  draw_group = function(data, panel_scales, coord) {
+    # Find the points for the line to go all the way around
+    data <- transform(data, xminv = x, xmaxv = x + violinwidth * (xmax - x))
+
+    # Make sure it's sorted properly to draw the outline
+    newdata <- rbind(
+      plyr::arrange(transform(data, x = xminv), y),
+      plyr::arrange(transform(data, x = xmaxv), -y)
+    )
+
+    # Close the polygon: set first and last point the same
+    # Needed for coord_polar and such
+    newdata <- rbind(newdata, newdata[1, ])
+
+    return(ggplot2:::ggname("geom_flat_violin", ggplot2::GeomPolygon$draw_panel(newdata, panel_scales, coord)))
+  },
+
+  draw_key = draw_key_polygon,
+  default_aes = aes(weight = 1, colour = "grey20", fill = "white", size = 0.5, alpha = NA, linetype = "solid"),
+  required_aes = c("x", "y")
+)