+++ R/utils.R
@@ -1,7 +1,7 @@
 #' @title Turn Null Into Character "NA"
 #' @description This function turns `NULL` in a list into character "NA", which
 #'   can be very useful when processing JSON data that has `null` values.
-#' @param x a list
+#' @param x A list
 #' @return If any element from the input list is NULL, they will be turned into character
 #'  "NA". Otherwise, return the original list.
 #' @examples \dontrun{
@@ -28,8 +28,29 @@ null2na <- function(x) {
 #'   **Note**: if the intention is to have the parsed data in a column such as
 #'   when parsing in [dplyr::mutate()], it is *highly* recommended to switch
 #'   to a `tibble` via [tibble::as_tibble()] *first*.
-#' @param x character vector of JSON
+#' @param x A character vector of JSON
 #' @export
 parse_json <- function(x) {
   return(purrr::map(x, jsonlite::fromJSON))
+}
+
+#' @title Invert list
+#' @description Inverts a (named) list such that the values become the keys and
+#'   the keys become the values.
+#' @param x A named list
+#' @return A list with values as keys and keys as values.
+#' @examples
+#' invert_list(list(x = c(1, 2), y = c(2, 3)))
+#' # returns: list(`1` = "x", `2` = c("x", "y"), `3` = "y")
+#' @author Mikhail Popov
+#' @export
+invert_list <- function(x) {
+  if (is.null(names(x))) stop("expecting input to be a named list")
+  new_fields <- Reduce(union, x)
+  if (length(new_fields) == 0) warning("inverted list will be empty")
+  names(new_fields) <- new_fields
+  return(lapply(new_fields, function(field) {
+    y <- purrr::map_lgl(x, ~ field %in% .x)
+    return(names(y)[y])
+  }))
 }